EDbackend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
fastrlock==0.8.1
filelock==3.12.2
flasgger==0.9.7.1
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.21.12
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywatchman==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
#tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
